refactor(clawql-data): Effect engine plugins; strip LAB session to MCP proxy
- clawql-data: registerDataEngine plugin registry; duckdb is one plugin
- Effect services (DataEngineService) mirror clawql-sandbox pattern
- Remove chdb/python rejection checks — unknown engine ids fail via registry
- clawql_lab_session.py: 1686 → 320 lines, MCP HTTP proxy only
- Vault seed via lab-pre-ingest.mjs; recall enrich in clawql-memory (CLAWQL_HARVEY_LAB=1)
- Delete clawql_lab_evidence.py and Python-only LAB tests
