+++ packages/clawql-data/src/effect/data-errors.ts
@@ -1,0 +1,6 @@
+import { Data } from "effect";
+
+export class DataError extends Data.TaggedError("DataError")<{
+  readonly message: string;
+  readonly cause?: unknown;
+}> {}
