+++ scripts/start-clawql-for-lab.sh
@@ -45,7 +45,7 @@ fi
 export CLAWQL_OBSIDIAN_VAULT_PATH="${VAULT_PATH}"
 export CLAWQL_ENABLE_MEMORY="${CLAWQL_ENABLE_MEMORY:-1}"
 export CLAWQL_ENABLE_DATA="${CLAWQL_ENABLE_DATA:-1}"
-export CLAWQL_DATA_ENGINE="${CLAWQL_DATA_ENGINE:-duckdb}"
+export CLAWQL_HARVEY_LAB="${CLAWQL_HARVEY_LAB:-1}"
 export CLAWQL_DATA_PATH="${CLAWQL_DATA_PATH:-${VAULT_PATH}/lab/matters.duckdb}"
 export CLAWQL_DATA_INGEST_ROOTS="${CLAWQL_DATA_INGEST_ROOTS:-/workspace:/tmp:${HOME}}"
 export CLAWQL_ONTOLOGY_DB="${CLAWQL_ONTOLOGY_DB:-1}"
